Revert "fix: Screen Reader: Workspace: 3-dot More button in Workspace container not focusable or operable"
